+++ tools/hrdf-db-importer/inc/HRDF_Parser/parse_betrieb.py
@@ -11,8 +11,6 @@ def import_db_betrieb(hrdf_path, db_path, db_schema_config):
     agency_row_items = parse_hrdf_betrieb(hrdf_path, fplan_agency_ids)
     truncate_and_load_table_records(db_path, 'agency', db_schema_config['tables']['agency'], agency_row_items)
 
-    print('')
-
 def fetch_agency_from_fplan(db_path):
     db_handle = sqlite3.connect(db_path)