+++ tools/hrdf-db-importer/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 15.Oct 2023
 - add support for journey information `A JY` from `INFOTEXT` - [PR #34](https://github.com/openTdataCH/OJP-Showcase/pull/34)
+- improve `GLEIS` parsing - [PR #35](https://github.com/openTdataCH/OJP-Showcase/pull/35)